ControlMessage exposed, reduced methods.
